backend/requirements.txt
ansicon==1.89.0
anyio==4.10.0
arrow==1.3.0
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
blessed==1.21.0
certifi==2025.8.3
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
constantly==23.10.4
cryptography==45.0.6
daphne==4.2.1
Django==5.2.5
django-cors-headers==4.7.0
django-picklefield==3.3
django-rq==3.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.0
filelock==3.18.0
fsspec==2025.7.0
gevent==25.5.1
greenlet==3.2.4
h11==0.16.0
hf-xet==1.1.7
httptools==0.6.4
huggingface-hub==0.34.4
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
Jinja2==3.1.6
jinxed==1.3.0
joblib==1.5.1
lxml==6.0.0
MarkupSafe==3.0.2
mpmath==1.3.0
msgpack==1.1.1
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
packaging==24.2
pillow==11.3.0
pinecone==7.3.0
pinecone-plugin-assistant==1.7.0
pinecone-plugin-interface==0.0.7
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
PyMuPDF==1.26.3
pyOpenSSL==25.1.0
pypdf==5.9.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
PyYAML==6.0.2
redis==6.4.0
regex==2025.7.34
requests==2.32.4
rq==2.4.1
rq-gevent-worker==0.1.6
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.15.3
sentence-transformers==5.1.0
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.4
tomli==2.2.1
torch==2.8.0
tqdm==4.67.1
transformers==4.55.0
triton==3.4.0
Twisted==25.5.0
txaio==25.6.1
types-python-dateutil==2.9.0.20250708
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
zope.event==5.1.1
zope.interface==7.2
